making final changes

+++ src/features/cart/cartSlice.js
@@ -5,7 +5,7 @@ const initialState = {
 
   // cart: [
   //   {
-  //     pizzaId: 23479,
+  //     pizzaId: 12,
   //     name: 'Mediterranean',
   //     quantity: 2,
   //     unitPrice: 16,
@@ -22,30 +22,26 @@ const cartSlice = createSlice({
       // payload = newItem
       state.cart.push(action.payload);
     },
-
     deleteItem(state, action) {
       // payload = pizzaId
       state.cart = state.cart.filter((item) => item.pizzaId !== action.payload);
     },
-
-    increaseItemQuntity(state, action) {
+    increaseItemQuantity(state, action) {
       // payload = pizzaId
       const item = state.cart.find((item) => item.pizzaId === action.payload);
-      item.quantity++;
-      item.totalPrice = item.unitPrice * item.quantity;
-    },
 
-    decreaseItemQuntity(state, action) {
+      item.quantity++;
+      item.totalPrice = item.quantity * item.unitPrice;
+    },
+    decreaseItemQuantity(state, action) {
       // payload = pizzaId
       const item = state.cart.find((item) => item.pizzaId === action.payload);
 
       item.quantity--;
-      item.totalPrice = item.unitPrice * item.quantity;
-      if (item.quantity === 0) {
-        cartSlice.caseReducers.deleteItem(state, action);
-      }
-    },
+      item.totalPrice = item.quantity * item.unitPrice;
 
+      if (item.quantity === 0) cartSlice.caseReducers.deleteItem(state, action);
+    },
     clearCart(state) {
       state.cart = [];
     },
@@ -55,8 +51,8 @@ const cartSlice = createSlice({
 export const {
   addItem,
   deleteItem,
-  increaseItemQuntity,
-  decreaseItemQuntity,
+  increaseItemQuantity,
+  decreaseItemQuantity,
   clearCart,
 } = cartSlice.actions;
 
@@ -72,3 +68,5 @@ export const getTotalCartPrice = (state) =>
 
 export const getCurrentQuantityById = (id) => (state) =>
   state.cart.cart.find((item) => item.pizzaId === id)?.quantity ?? 0;
+
+// 'reselect'
